Create simple positive and negative test for model property isNull to check if all values of model are null; moved presenter to correct package; updated gradle build dependencies
